JWT Added

+++ src/main/java/com/example/BookHeaven/service/UserService.java
@@ -1,39 +1,50 @@
 package com.example.BookHeaven.service;
 
-import com.example.BookHeaven.model.User;
-import com.example.BookHeaven.repository.UserRepository;
+import java.util.List;
 
-import org.springframework.beans.factory.annotation.Autowired;
-//import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.security.authentication.AuthenticationManager;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.AuthenticationException;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.stereotype.Service;
 
-
-import java.util.List;
+import com.example.BookHeaven.Utils.JwtUtil;
+import com.example.BookHeaven.model.User;
+import com.example.BookHeaven.repository.UserRepository;
 
 @Service
 public class UserService {
 
-    @Autowired
-    private UserRepository userRepository;
-    
-//    @Autowired
-//    private PasswordEncoder passwordEncoder;
+    // @Autowired
+    private final UserRepository userRepository;
+    private final AuthenticationManager authenticationManager;
+    private final JwtUtil jwtUtil;
+    private final UserDetailsService userDetailsService;
+
+    UserService(UserRepository userRepository, AuthenticationManager authenticationManager, JwtUtil jwtUtil,
+            UserDetailsService userDetailsService) {
+        this.userRepository = userRepository;
+        this.authenticationManager = authenticationManager;
+        this.jwtUtil = jwtUtil;
+        this.userDetailsService = userDetailsService;
+    }
 
     public List<User> getAllUsers() {
         return userRepository.findAll();
     }
-    
+
     public User getUserByEmail(String email) {
         return userRepository.findByEmail(email);
     }
-    
+
     public User getUserById(String id) {
         return userRepository.findById(id).orElse(null);
     }
 
     public User createUser(User user) {
-    	try {
-//    		user.setPassword(passwordEncoder.encode(user.getPassword()));
+        try {
+            // user.setPassword(passwordEncoder.encode(user.getPassword()));
             return userRepository.save(user);
         } catch (Exception e) {
             // Handle duplicate email exception
@@ -46,7 +57,17 @@ public class UserService {
         return userRepository.save(user);
     }
 
-    public void deleteUser(String id) {
-        userRepository.deleteById(id);
+    public User deleteUser(String id) {
+        User user = userRepository.findById(id).orElse(null);
+        if (user != null) {
+            userRepository.deleteById(id);
+        }
+        return user;
+    }
+
+    public String authenticate(String email, String password) throws AuthenticationException {
+        authenticationManager.authenticate(new UsernamePasswordAuthenticationToken(email, password));
+        final UserDetails userDetails = userDetailsService.loadUserByUsername(email);
+        return jwtUtil.generateToken(userDetails.getUsername());
     }
 }
